+++ CHANGELOG.md
@@ -1,11 +1,21 @@
+---
+
+### 0.0.7 : Added ( SuperPositioned() ) Widget
 ### 0.0.6 : Added ( Aligner() ) class
-
 ### 0.0.5 : Upgrades
-
 ### 0.0.4 : Added ( Borderers() ) class
-
 ### 0.0.3 : Added ( screenShortestSide() )
-
 ### 0.0.2 : ScreenHeight method no deducts screen vertical paddings by default
-
 ### 0.0.1 : Separated Scale class in this package
+
+---
+
+## Included Classes :-
+
+* Scale(); // handles MediaQuery methods
+* Aligners(); // handles Alignment methods
+* Borderers(); // handles BorderRadius and corners
+
+* SuperPositioned(); // handles Positioned Widget in
+
+---